fl-node/pyproject.toml
[project]
name = "fl-node"
version = "0.1.0"
description = "Federated Learning Node for the Cancer Screening Platform"
authors = [
    {name = "gulsah"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "flwr==1.20.0",
    "numpy==2.3.2",
    "requests==2.32.4",
    "python-dotenv==1.1.1",
    "cryptography==44.0.1",
    "monai==1.5.0",
    "timm==1.0.19",
    "tenseal==0.3.16",
    "dvc==3.61.0",
    "pydicom==3.0.1",
    "Pillow==11.3.0",
    "mlflow==3.2.0",
    "opacus==1.5.4",
    "pytest==8.4.1",
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"

[tool.poetry]
packages = [{include = "src", from = "."}]
package-mode = false



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-app = "poetry.core.masonry.api"